+++ docs/README.md
@@ -130,6 +130,10 @@
 
 - Orders
   - [ ] 음식의 총 개수가 21개 이상일 때 예외를 발생시킨다.
-  - [ ] 메뉴판에 없는 메뉴를 주문하면 예외를 발생시킨다.
   - [ ] 중복 메뉴를 입력했을 때 예외를 발생시킨다.
   - [ ] 음료만 주문했을 때 예외를 발생시킨다.
+
+<br/>
+
+- OrderItem
+  - [x] 메뉴판에 없는 메뉴를 주문하면 예외를 발생시킨다.
